# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticationVerificationDto": {
        "description": "Pydantic transport contract for a completed passkey authentication.",
        "properties": {
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          }
        },
        "required": ["csrf_token"],
        "title": "AuthenticationVerificationDto",
        "type": "object"
      },
...
      "ChallengeDto": {
        "properties": {
          "challenge_id": {
            "title": "Challenge Id",
            "type": "string"
          },
          "options": {
            "additionalProperties": true,
            "title": "Options",
            "type": "object"
          }
        },
        "required": ["challenge_id", "options"],
        "title": "ChallengeDto",
        "type": "object"
      },
...
      "CredentialRequest": {
        "properties": {
          "challenge_id": {
            "title": "Challenge Id",
            "type": "string"
          },
          "credential": {
            "additionalProperties": true,
            "title": "Credential",
            "type": "object"
          }
        },
        "required": ["challenge_id", "credential"],
        "title": "CredentialRequest",
        "type": "object"
      },
...
      "OpaqueRequestErrorDto": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "const": "request is not permitted",
            "default": "request is not permitted",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "OpaqueRequestErrorDto",
        "type": "object"
      },
...
      "ReauthenticationVerificationDto": {
        "description": "Pydantic transport contract for a completed recent reauthentication.",
        "properties": {
          "credential_id": {
            "title": "Credential Id",
            "type": "string"
          }
        },
        "required": ["credential_id"],
        "title": "ReauthenticationVerificationDto",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Stock Profiler",
    "version": "0.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/passkeys/authentication/options": {
      "post": {
        "operationId": "authentication_options_api_v1_auth_passkeys_authentication_options_post",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "No enrolled passkey is available"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Authentication Options"
      }
    },
    "/api/v1/auth/passkeys/authentication/verify": {
      "post": {
        "operationId": "authentication_verify_api_v1_auth_passkeys_authentication_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticationVerificationDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication verification failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Authentication Verify"
      }
    },
    "/api/v1/auth/passkeys/reauthentication/options": {
      "post": {
        "operationId": "reauthentication_options_api_v1_auth_passkeys_reauthentication_options_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-stock_profiler_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Stock Profiler Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Recent authenticated session required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Reauthentication Options"
      }
    },
    "/api/v1/auth/passkeys/reauthentication/verify": {
      "post": {
        "operationId": "reauthentication_verify_api_v1_auth_passkeys_reauthentication_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-stock_profiler_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Stock Profiler Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReauthenticationVerificationDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Reauthentication verification failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Reauthentication Verify"
      }
    },
    "/api/v1/auth/passkeys/registration/options": {
      "post": {
        "operationId": "registration_options_api_v1_auth_passkeys_registration_options_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Host-Console-Grant",
            "required": true,
            "schema": {
              "title": "X-Host-Console-Grant",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Host console grant required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Registration Options"
      }
    },
    "/api/v1/auth/passkeys/registration/verify": {
      "post": {
        "operationId": "registration_verify_api_v1_auth_passkeys_registration_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Registration verification failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Registration Verify"
      }
    },
    "/api/v1/auth/session": {
      "delete": {
        "operationId": "logout_api_v1_auth_session_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-stock_profiler_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Stock Profiler Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "description": "Origin, CSRF token, or session was invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Logout"
      }
    },
    "/api/v1/auth/session/refresh": {
      "post": {
        "operationId": "refresh_session_api_v1_auth_session_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-stock_profiler_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Stock Profiler Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRefreshDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "Origin, CSRF token, or session was invalid"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Refresh Session"
      }
    },
    "/api/v1/diagnostics/safety-capabilities": {
      "get": {
        "description": "Report static safety boundaries without exposing a product action.",
        "operationId": "safety_capabilities_diagnostic_api_v1_diagnostics_safety_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafetyCapabilitiesDiagnosticDto"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpaqueRequestErrorDto"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Safety Capabilities Diagnostic"
      }
    },
...
  }
}